# docs/reference/openapi.json
{
  "components": {
    "schemas": {
      "ApprovalDecisionRequest": {
        "description": "A person's decision on an approval a run is waiting for.",
        "properties": {
          "approver": {
            "description": "Who decided",
            "minLength": 1,
            "title": "Approver",
            "type": "string"
          },
          "arguments": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Approve this edited call instead of the one asked for",
            "title": "Arguments"
          },
          "decision": {
            "enum": [
              "approve",
              "deny"
            ],
            "title": "Decision",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reason; a denial's note reaches the model",
            "title": "Note"
          }
        },
        "required": [
          "decision",
          "approver"
        ],
        "title": "ApprovalDecisionRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Standard error response model.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional error details",
            "title": "Detail"
          },
          "error": {
            "description": "Error type",
            "title": "Error",
            "type": "string"
          },
          "message": {
            "description": "Error message",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunRequest": {
        "description": "Request model for agent run endpoint.",
        "properties": {
          "query": {
            "description": "The query/prompt for the agent",
            "title": "Query",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional session ID for conversation continuity",
            "title": "Session Id"
          }
        },
        "required": [
          "query"
        ],
        "title": "RunRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The HTTP API of an agent served with `omniserve run --agent agent.py`: runs, approvals, budgets, background tasks, sessions and the evidence of every run.",
    "title": "OmniServe",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/run": {
      "post": {
        "description": "Run the agent with a query and stream SSE events.",
        "operationId": "run_agent_sse_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "SSE stream of agent events"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Run agent (SSE streaming)",
        "tags": [
          "Runs"
        ]
      }
    },
    "/run/sync": {
      "post": {
        "description": "Run the agent with a query and return a JSON response.",
        "operationId": "run_agent_sync_run_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Run agent (synchronous)",
        "tags": [
          "Runs"
        ]
      }
    },
    "/runs": {
      "get": {
        "description": "Runs by status and session, oldest first: what waits for a person (`awaiting_approval`, `awaiting_budget`), what a crash left `running`, what finished. Each as `GET /runs/{run_id}` returns it; no saved conversation.",
        "operationId": "list_runs_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Runs Runs Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List runs",
        "tags": [
          "Runs"
        ]
      }
    },
    "/runs/{run_id}": {
      "get": {
        "description": "A run's durable record: status, step, usage, trace IDs, tool call states, and approvals. Its saved conversation is not returned.",
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Run Runs  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a run",
        "tags": [
          "Runs"
        ]
      }
    },
    "/runs/{run_id}/approvals/{approval_id}": {
      "post": {
        "description": "Approve or deny an approval a paused run is waiting for.",
        "operationId": "decide_approval_runs__run_id__approvals__approval_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Decide Approval Runs  Run Id  Approvals  Approval Id  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide an approval",
        "tags": [
          "Runs"
        ]
      }
    },
    "/runs/{run_id}/budget": {
      "get": {
        "description": "Every budget covering the run \u2014 its own, its session's, its agent's, the application's \u2014 with the limit, what a person granted, what is spent and reserved, and the ledger key; and `requests`, each time the run ran out (pending while it waits, then granted or denied).",
        "operationId": "read_budgets_runs__run_id__budget_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Budgets Runs  Run Id  Budget Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a run's budgets",
        "tags": [
          "Runs"
        ]
      },
      "post": {
        "description": "Grant or refuse the budget a waiting run ran out of. A grant is a recorded exception to that one budget, with the name of whoever made it; the policy is unchanged. Resume the run afterwards.",
        "operationId": "decide_budget_runs__run_id__budget_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Decide Budget Runs  Run Id  Budget Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide a budget",
        "tags": [
          "Runs"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "omniserve run, locally",
      "url": "http://localhost:8000"
    }
  ]
}